+++ tests/AbstractTestCase.php
@@ -1,41 +1,76 @@
 <?php
 
+declare(strict_types = 1);
+
 namespace AvtoDev\FirebaseNotificationsChannel\Tests;
 
 use GuzzleHttp\Client;
 use GuzzleHttp\HandlerStack;
 use GuzzleHttp\Handler\MockHandler;
+use Illuminate\Contracts\Console\Kernel;
+use Illuminate\Foundation\Testing\TestCase;
 use AvtoDev\FirebaseNotificationsChannel\FcmClient;
-use AvtoDev\DevTools\Tests\PHPUnit\AbstractLaravelTestCase;
+use AvtoDev\FirebaseNotificationsChannel\FcmChannel;
+use AvtoDev\FirebaseNotificationsChannel\ServiceProvider;
 
-abstract class AbstractTestCase extends AbstractLaravelTestCase
+abstract class AbstractTestCase extends TestCase
 {
     /**
      * @var MockHandler
      */
     protected $mock_handler;
 
-    public function setUp()
+    /**
+     * @var Client
+     */
+    protected $http_client;
+
+    /**
+     * {@inheritdoc}
+     *
+     * @throws \InvalidArgumentException
+     */
+    public function setUp(): void
     {
         parent::setUp();
 
-        $this->app->bind(FcmClient::class, function () {
-            $this->mock_handler = new MockHandler;
+        $this->mock_handler = new MockHandler;
 
-            $handler = HandlerStack::create($this->mock_handler);
+        $handler = HandlerStack::create($this->mock_handler);
 
-            $http_client = new Client(['handler' => $handler]);
+        $this->http_client = $http_client = new Client(['handler' => $handler]);
 
+        $binding = static function () use ($http_client) {
             return new FcmClient(
                 $http_client,
                 'https://fcm.googleapis.com/v1/projects/' . config('fcm.project_id') . '/messages:send'
             );
-        });
+        };
+
+        $this->app->bind(FcmClient::class, $binding);
+
+        $this->app
+            ->when(FcmChannel::class)
+            ->needs(FcmClient::class)
+            ->give($binding);
     }
 
-    public function tearDown()
+    /**
+     * Creates the application.
+     *
+     * @return \Illuminate\Foundation\Application
+     */
+    public function createApplication()
     {
-        \Mockery::close();
-        parent::tearDown();
+        /** @var \Illuminate\Foundation\Application $app */
+        $app = require __DIR__ . '/../vendor/laravel/laravel/bootstrap/app.php';
+
+        $app->make(Kernel::class)->bootstrap();
+
+        $app->make('config')->set('services', require __DIR__ . '/config/services.php');
+
+        $app->register(ServiceProvider::class);
+
+        return $app;
     }
 }