+++ tests/Bootstrapper/bootstrap.php
@@ -1,3 +1,0 @@
-<?php
-
-new AvtoDev\FirebaseNotificationsChannel\Tests\Bootstrapper\Bootstrapper;