+++ tests/PlatformSettings/AbstractPlatformSettingsTest.php
@@ -1,7 +1,10 @@
 <?php
 
+declare(strict_types = 1);
+
 namespace AvtoDev\FirebaseNotificationsChannel\Tests\PlatformSettings;
 
+use Illuminate\Support\Arr;
 use Illuminate\Support\Str;
 use Illuminate\Contracts\Support\Arrayable;
 use AvtoDev\FirebaseNotificationsChannel\Tests\AbstractTestCase;
@@ -9,28 +12,28 @@ use AvtoDev\FirebaseNotificationsChannel\Tests\AbstractTestCase;
 abstract class AbstractPlatformSettingsTest extends AbstractTestCase
 {
     /**
+     * Must contains array of array.
+     *
+     * [
+     *   [$property, $array_path, $value]
+     * ]
+     *
      * @return array
      */
-    abstract public function dataProvider();
+    abstract public function dataProvider(): array;
 
     /**
-     * @param $property
-     * @param $array_path
-     * @param $value
-     *
-     * @throws \ReflectionException
      * @throws \SebastianBergmann\RecursionContext\InvalidArgumentException
-     *
-     * @dataProvider dataProvider
      */
-    public function testSetters($property, $array_path, $value)
+    public function testSetters(): void
     {
         $platform_settings = $this->getPlatformSetting();
 
-        $platform_settings->{'set' . Str::camel($property)}($value);
+        foreach ($this->dataProvider() as [$property, $array_path, $value]) {
+            $platform_settings->{'set' . Str::camel($property)}($value);
 
-        static::assertEquals($value, static::getProperty($platform_settings, $property));
-        static::assertEquals($value, array_get($platform_settings->toArray(), $array_path));
+            $this->assertEquals($value, Arr::get($platform_settings->toArray(), $array_path));
+        }
     }
 
     /**