+++ tests/config/services.php
@@ -26,7 +26,7 @@ return [
          |
          */
 
-        'driver' => env('FCM_DRIVER', ''),
+        'driver' => env('FCM_DRIVER', 'config'),
 
         /*
          |---------------------------------------------------------------------
@@ -56,21 +56,41 @@ return [
                 */
 
                 'credentials' => [
-                    'private_key_id'              => env('FCM_CREDENTIALS_PRIVATE_KEY_ID',
-                        'da80b3bbceaa554442ad67e6be361a66'),
-                    'private_key'                 => env('FCM_CREDENTIALS_PRIVATE_KEY',
-                        '-----BEGIN PRIVATE KEY-----\n...\n-----END PRIVATE KEY-----\n'),
-                    'client_email'                => env('FCM_CREDENTIALS_CLIENT_EMAIL',
-                        'firebase-adminsdk-mwax6@test.iam.gserviceaccount.com'),
-                    'client_id'                   => env('FCM_CREDENTIALS_CLIENT_ID', '22021520333507180281'),
-                    'auth_uri'                    => env('FCM_CREDENTIALS_AUTH_URI',
-                        'https://accounts.google.com/o/oauth2/auth'),
-                    'token_uri'                   => env('FCM_CREDENTIALS_TOKEN_URI',
-                        'https://oauth2.googleapis.com/token'),
-                    'auth_provider_x509_cert_url' => env('FCM_CREDENTIALS_AUTH_PROVIDER_CERT',
-                        'https://www.googleapis.com/oauth2/v1/certs'),
-                    'client_x509_cert_url'        => env('FCM_CREDENTIALS_CLIENT_CERT',
-                        'https://www.googleapis.com/robot/v1/metadata/x509/firebase-adminsdk-mwax6%40test.iam.gserviceaccount.com'),
+                    'type'       => 'service_account',
+                    'project_id' => env('FCM_CREDENTIALS_PROJECT_ID', 'some_id'),
+
+                    'private_key_id'              => env(
+                        'FCM_CREDENTIALS_PRIVATE_KEY_ID',
+                        'da80b3bbceaa554442ad67e6be361a66'
+                    ),
+                    'private_key'                 => env(
+                        'FCM_CREDENTIALS_PRIVATE_KEY',
+                        '-----BEGIN PRIVATE KEY-----\n...\n-----END PRIVATE KEY-----\n'
+                    ),
+                    'client_email'                => env(
+                        'FCM_CREDENTIALS_CLIENT_EMAIL',
+                        'firebase-adminsdk-mwax6@test.iam.gserviceaccount.com'
+                    ),
+                    'client_id'                   => env(
+                        'FCM_CREDENTIALS_CLIENT_ID',
+                        '22021520333507180281'
+                    ),
+                    'auth_uri'                    => env(
+                        'FCM_CREDENTIALS_AUTH_URI',
+                        'https://accounts.google.com/o/oauth2/auth'
+                    ),
+                    'token_uri'                   => env(
+                        'FCM_CREDENTIALS_TOKEN_URI',
+                        'https://oauth2.googleapis.com/token'
+                    ),
+                    'auth_provider_x509_cert_url' => env(
+                        'FCM_CREDENTIALS_AUTH_PROVIDER_CERT',
+                        'https://www.googleapis.com/oauth2/v1/certs'
+                    ),
+                    'client_x509_cert_url'        => env(
+                        'FCM_CREDENTIALS_CLIENT_CERT',
+                        'https://www.googleapis.com/robot/v1/metadata/x509/firebase-adminsdk-mwax6%40test.iam.gserviceaccount.com'
+                    ),
                 ],
             ],
         ],